Add test-release.json
This file contains the list of pre-released cards that are coming out, as well as the database. When these cards are released, this file is updated. You can see the original here:
https://cdn02.moecube.com:444/ygopro-super-pre/data/test-release.json
What we need to do is download the original and translate the cards that are coming out. In the future, Sherry will use this file to display these cards directly from the application.